舊的程式檔案/GoogleAPI-爬蟲/poetry-finalproject/pyproject.toml
[tool.poetry]
name = "poetry-finalproject"
version = "0.1.0"
description = ""
authors = ["itsnataliekkk <10946001@ntub.edu.tw>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
